refactor: rename 'config' to 'configuration'
