+++ src/configuration/external/adapt.ts
