+++ src/configuration/external/common.ts
