+++ src/configuration/external/index.ts
