+++ src/configuration/standard/index.ts
