+++ src/configuration/standard/nsci.ts
