+++ src/reporting/reporters/post-pipeline/console/index.ts
@@ -4,7 +4,7 @@ import ms from "pretty-ms";
 
 import { consolePrinter } from "../../../../../lib/console-printer/index.js";
 import { OutcomePayloadFromPipelineChecks } from "../../../../analysis/interpret.js";
-import { Nsci } from "../../../../config/standard/index.js";
+import { Nsci } from "../../../../configuration/standard/index.js";
 import { Reporter } from "../../reporter.js";
 
 import { printPipelineOutcome } from "./outcome.js";
