+++ src/reporting/reporters/post-pipeline/console/outcome.ts
@@ -5,7 +5,7 @@ import {
   consolePrinter
 } from "../../../../../lib/console-printer/index.js";
 import type { CompactedScannerPayload } from "../../../../analysis";
-import { Warnings } from "../../../../config/standard/nsci.js";
+import { Warnings } from "../../../../configuration/standard/nsci.js";
 import { pipeline } from "../../../index.js";
 
 import { printWarnOrError } from "./util.js";
