+++ src/reporting/reporters/post-pipeline/console/warnings.ts
@@ -1,7 +1,7 @@
 import pluralize from "pluralize";
 
 import { consolePrinter } from "../../../../../lib/console-printer/index.js";
-import { Warnings } from "../../../../config/standard/nsci.js";
+import { Warnings } from "../../../../configuration/standard/nsci.js";
 import type { DependencyWarning } from "../../../../types";
 
 import { printWarnOrError } from "./util.js";
